PostgreSQL with Knex: Populating The Populations

+++ postgres_with_knex/labs/knex_labs/db/migrations/20230302013212_create_populations_table.js
@@ -9,7 +9,7 @@ exports.up = function(knex) {
   return knex.schema.createTable('populations', table => {
     table.bigIncrements('id');
     table.integer('year');
-    table.integer('quantity');
+    table.bigInteger('quantity');
     table.integer('country_id');
     table.foreign('country_id').references('countries.id');
     // foreign key can only be declared on an existing column.
